Moved test doubles into their own classes
