+++ core/src/test/java/com/novoda/merlin/service/MerlinServiceShould.java
@@ -182,64 +182,4 @@ public class MerlinServiceShould {
         return new ConnectivityChangeEvent(isConnected, "info", "reason");
     }
 
-    private static class TestDoubleMerlinService extends MerlinService {
-
-        private final PackageManager packageManager;
-        private final ComponentName receiver;
-
-        TestDoubleMerlinService(PackageManager packageManager, ComponentName receiver) {
-            this.packageManager = packageManager;
-            this.receiver = receiver;
-        }
-
-        @Override
-        public PackageManager getPackageManager() {
-            return packageManager;
-        }
-
-        @Override
-        protected ComponentName connectivityReceiverComponent() {
-            return receiver;
-        }
-    }
-
-    private static class TestDoubleMerlinServiceWithStubbedBuilders extends MerlinService {
-
-        private final ContextWrapper context;
-        private final NetworkStatusRetriever networkStatusRetriever;
-        private final HostPinger defaultHostPinger;
-        private final HostPinger customHostPinger;
-
-        TestDoubleMerlinServiceWithStubbedBuilders(
-                ContextWrapper context,
-                NetworkStatusRetriever networkStatusRetriever,
-                HostPinger defaultHostPinger,
-                HostPinger customHostPinger
-        ) {
-            this.context = context;
-            this.networkStatusRetriever = networkStatusRetriever;
-            this.defaultHostPinger = defaultHostPinger;
-            this.customHostPinger = customHostPinger;
-        }
-
-        @Override
-        public ContextWrapper getApplicationContext() {
-            return context;
-        }
-
-        @Override
-        protected NetworkStatusRetriever buildNetworkStatusRetriever() {
-            return networkStatusRetriever;
-        }
-
-        @Override
-        protected HostPinger buildDefaultHostPinger() {
-            return defaultHostPinger;
-        }
-
-        @Override
-        protected HostPinger buildHostPinger(String hostName, ResponseCodeValidator validator) {
-            return customHostPinger;
-        }
-    }
 }
